requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
aiosqlite
amqtt
